graph 폴더 refactor (#29)
