+++ app/graph/nodes/answer/smalltalk_node.py
