+++ app/graph/nodes/input/stt_node.py
