+++ app/graph/nodes/tool/struct_db_node.py
